feat(cli): update warning description according to installed or not
